refactor: 친구 요청 관련 코드를 friendrequest → friend 패키지로 통합
