+++ src/main/java/life/walkit/server/friend/repository/FriendRequestRepository.java
@@ -1,7 +1,7 @@
-package life.walkit.server.friendrequest.repository;
+package life.walkit.server.friend.repository;
 
-import life.walkit.server.friendrequest.entity.FriendRequest;
-import life.walkit.server.friendrequest.enums.FriendRequestStatus;
+import life.walkit.server.friend.entity.FriendRequest;
+import life.walkit.server.friend.enums.FriendRequestStatus;
 import life.walkit.server.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -13,4 +13,7 @@ public interface FriendRequestRepository extends JpaRepository<FriendRequest, Lo
 
     // 친구 요청 시 예외처리 로직
     boolean existsBySenderAndReceiver(Member sender, Member receiver);
+
+    // 내가 보낸 친구 요청을 조회하기 위한 메서드
+    List<FriendRequest> findBySender(Member sender);
 }
