+++ src/test/java/life/walkit/server/global/factory/GlobalTestFactory.java
@@ -1,7 +1,7 @@
 package life.walkit.server.global.factory;
 
 import life.walkit.server.friend.entity.Friend;
-import life.walkit.server.friendrequest.entity.FriendRequest;
+import life.walkit.server.friend.entity.FriendRequest;
 import life.walkit.server.member.entity.Member;
 import life.walkit.server.member.entity.enums.MemberRole;
 import life.walkit.server.member.entity.enums.MemberStatus;
